requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
fake-useragent==1.2.1
python-dotenv==1.0.0
openpyxl==3.1.2
lxml=5.3.0